refactor: inline discovery/registry.rs into discovery/mod.rs
